+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Schema documentation available at https://oxid-esales.github.io/graphql-storefront-module
+- Classes:
+  - `OxidEsales\GraphQL\Storefront\Basket\DataType\BasketByTitleAndUserIdFilterList`
+  - `OxidEsales\GraphQL\Storefront\Basket\DataType\Sorting`
+- Method `OxidEsales\GraphQL\Storefront\Basket\Infrastructure\Repository::basketExistsByTitleAndUserId()`
 
 ### Fixed
 - Remove ``final`` statement from shop extending classes [PR-3](https://github.com/OXID-eSales/graphql-storefront-module/pull/3).